+++ src/components/Main/Metrics/MetricsEmployee/MetricsEmployee.jsx
@@ -5,36 +5,63 @@ import Office from "./Office/Office.jsx";
 import BarChartComponent from "./BarChart/BarChart.jsx";
 import Workstation from "./Workstation/Workstation.jsx";
 import PieChartComponent from "./PieChart/PieChart.jsx";
-import { apiCallForMetrics, MetricsContext } from "../../../../store/metrics-context.jsx";
-import { $MetricsEmployee } from "./MetricsEmployee.styles.jsx";
+import {
+  apiCallForMetrics,
+  MetricsContext,
+} from "../../../../store/metrics-context.jsx";
+import { $MetricsEmployee, $ErrorLabel } from "./MetricsEmployee.styles.jsx";
 import { apiCallToGetEmployeeCarbonFootprint } from "../../../../api/Api.jsx";
 
 export default function MetricsEmployee() {
   const { currentStep } = useContext(MetricsContext);
   const [employeeMetric, setEmployeeMetric] = useState({});
+  const [error, setError] = useState(null);
 
   useEffect(() => {
     async function fetchData() {
-      try {
-        await apiCallForMetrics(currentStep).then(resData => {
-          const { id, corporateKey, email, name, surname, role, carbonLimit, location, office, workstation } = resData;
-          setEmployeeMetric((_prevMetrics) => ({ id, corporateKey, email, name, surname, role, carbonLimit, location, office, workstation }));
+      await apiCallForMetrics(currentStep)
+        .then((resData) => {
+          const {
+            id,
+            corporateKey,
+            email,
+            name,
+            surname,
+            role,
+            carbonLimit,
+            location,
+            office,
+            workstation,
+          } = resData;
+          setEmployeeMetric((_prevMetrics) => ({
+            id,
+            corporateKey,
+            email,
+            name,
+            surname,
+            role,
+            carbonLimit,
+            location,
+            office,
+            workstation,
+          }));
+        })
+        .catch((error) => {
+          setError(error);
         });
-      } catch (error) {
-        console.log(error);
-        //TODO
-      }
     }
 
-    fetchData()
+    fetchData();
   }, []);
 
   const carbonFootprint = useMemo(() => {
     return apiCallToGetEmployeeCarbonFootprint(0);
+    // TO DO
   }, [employeeMetric]);
 
   function getEmployeeData() {
-    const { corporateKey, email, name, surname, role, location } = employeeMetric;
+    const { corporateKey, email, name, surname, role, location } =
+      employeeMetric;
     return { corporateKey, email, name, surname, role, location };
   }
 
@@ -49,13 +76,26 @@ export default function MetricsEmployee() {
   }
 
   return (
-    <$MetricsEmployee>
-      <Employee employee={getEmployeeData()} />
-      <Office office={getOfficeData()} />
-      <Workstation workstation={getWorkstationData()} />
-      <Carbon employeeId={employeeMetric.id} carbonFootprint={carbonFootprint} carbonLimit={employeeMetric.carbonLimit} />
-      <BarChartComponent carbonFootprint={carbonFootprint} carbonLimit={employeeMetric.carbonLimit} />
-      <PieChartComponent carbonFootprint={carbonFootprint} />
-    </$MetricsEmployee>
+    <>
+      {error === null ? (
+        <$MetricsEmployee>
+          <Employee employee={getEmployeeData()} />
+          <Office office={getOfficeData()} />
+          <Workstation workstation={getWorkstationData()} />
+          <Carbon
+            employeeId={employeeMetric.id}
+            carbonFootprint={carbonFootprint}
+            carbonLimit={employeeMetric.carbonLimit}
+          />
+          <BarChartComponent
+            carbonFootprint={carbonFootprint}
+            carbonLimit={employeeMetric.carbonLimit}
+          />
+          <PieChartComponent carbonFootprint={carbonFootprint} />
+        </$MetricsEmployee>
+      ) : (
+        <$ErrorLabel>{error.message}</$ErrorLabel>
+      )}
+    </>
   );
 }