+++ src/components/Main/Selector/SelectorForm/Select/SelectValue.jsx
@@ -3,7 +3,7 @@ import { SelectorContext } from "../../../../../store/selector-context.jsx";
 import $Select from "./Select.styles.jsx";
 import { apiCallToGetFilterValues } from "../../../../../api/Api.jsx";
 
-export default function SelectValue() {
+export default function SelectValue({ errorHandler }) {
   const [filterValues, setFilterValues] = useState([]);
   const [firstOption, setFirstOption] = useState("");
   const { tempWhereCriteria, setTempWhereCriteria } =
@@ -11,36 +11,44 @@ export default function SelectValue() {
 
   useEffect(() => {
     async function fetchData() {
-      try {
-        await apiCallToGetFilterValues(tempWhereCriteria.key).then(
-          (resData) => {
-            setFilterValues((_prevFilterValues) => resData);
-            setTempWhereCriteria({
-              ...tempWhereCriteria,
-              value: resData ? resData[0].name : null,
-              id: resData ? resData[0].id : null,
-            });
-            setFirstOption(resData ? resData[0].name : "");
-          }
-        );
-      } catch (error) {
-        //TODO
-      }
+      await apiCallToGetFilterValues(tempWhereCriteria.key)
+        .then((resData) => {
+          setFilterValues((_prevFilterValues) => resData);
+          setTempWhereCriteria({
+            ...tempWhereCriteria,
+            value: resData ? resData[0].name : null,
+            id: resData ? resData[0].id : null,
+          });
+          setFirstOption(resData ? resData[0].name : "");
+        })
+        .catch((error) => {
+          console.error(error);
+          errorHandler(error)});
     }
-    fetchData();
-  }, [tempWhereCriteria.key]);
+    tempWhereCriteria.key && fetchData();
+  }, [tempWhereCriteria]);
 
   function handleWhereCriteriaChange(event) {
     setFirstOption(event.target.value);
     const selectedOption = event.target.options[event.target.selectedIndex];
-    const selectedKey = selectedOption.getAttribute('id');
-    setTempWhereCriteria({ ...tempWhereCriteria, value: event.target.value, id: selectedKey });
+    const selectedKey = selectedOption.getAttribute("id");
+    setTempWhereCriteria({
+      ...tempWhereCriteria,
+      value: event.target.value,
+      id: selectedKey,
+    });
   }
 
   return (
-    <$Select onChange={handleWhereCriteriaChange} value={firstOption} disabled={filterValues.length === 0}>
+    <$Select
+      onChange={handleWhereCriteriaChange}
+      value={firstOption}
+      disabled={filterValues.length === 0}
+    >
       {filterValues.map((filterValue, index) => (
-        <option id={filterValue.id} key={index}>{filterValue.name}</option>
+        <option id={filterValue.id} key={index}>
+          {filterValue.name}
+        </option>
       ))}
     </$Select>
   );